+++ src/components/forms/farmer-seed/edit/index.tsx
@@ -1,0 +1,5 @@
+const FarmerSeedEdit = () => {
+  return <div>FarmerSeedEdit</div>;
+};
+
+export default FarmerSeedEdit;