+++ src/routes/store-admin/_authenticated/farmer-seed/edit/index.tsx
@@ -1,0 +1,8 @@
+import { createFileRoute } from '@tanstack/react-router';
+import FarmerSeedEdit from '@/components/forms/farmer-seed/edit';
+
+export const Route = createFileRoute(
+  '/store-admin/_authenticated/farmer-seed/edit/'
+)({
+  component: FarmerSeedEdit,
+});